backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# WebSocket
python-socketio==5.11.0
python-engineio==4.9.0

# Serial Communication
pyserial==3.5
pyserial-asyncio==0.6

# Arduino CLI Wrapper
subprocess-run==0.1.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# HTTP Client
httpx==0.26.0
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# CORS
fastapi-cors==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
